+++ peer-crypto/src/main/java/com/exscudo/peer/core/crypto/IFormatter.java
@@ -1,4 +1,4 @@
-package com.exscudo.peer.core.common;
+package com.exscudo.peer.core.crypto;
 
 import java.util.Map;